+++ lib/runners/schema/result.rb
@@ -21,7 +21,12 @@ module Runners
       let :missing_file_failure, object(guid: string, timestamp: string, type: literal("missing_files"), files: array(string))
       let :error, object(guid: string, timestamp: string, type: literal("error"), class: string, backtrace: array(string), inspect: string)
 
-      let :envelope, object(result: enum(success, failure, missing_file_failure, error), warnings: array(warning), ci_config: any?)
+      let :envelope, object(
+        result: enum(success, failure, missing_file_failure, error),
+        warnings: array(warning),
+        ci_config: any?,
+        version: string,
+      )
     end
   end
 end